feat(mlop-2760): moving reader_metadata to a new module
