+++ butterfree/metadata/reader_metadata.py
